+++ crates/ewwii/src/widgets/build_widget.rs
@@ -50,9 +50,9 @@ fn build_gtk_widget_from_node(
         WidgetNode::ToolTip { props, children } => {
             build_tooltip(props, children, widget_reg)?.upcast()
         }
-        // WidgetNode::CircularProgress { props } => {
-        //     build_circular_progress_bar(props, widget_reg)?.upcast()
-        // }
+        WidgetNode::CircularProgress { props } => {
+            build_circular_progress_bar(props, widget_reg)?.upcast()
+        }
         // WidgetNode::Graph { props } => build_graph(props, widget_reg)?.upcast(),
         // WidgetNode::Transform { props } => build_transform(props, widget_reg)?.upcast(),
         WidgetNode::Slider { props } => build_gtk_scale(props, widget_reg)?.upcast(),